Commit days ago long to int

+++ src/main/java/com/alorma/github/sdk/bean/dto/response/Commit.java
@@ -10,7 +10,7 @@ public class Commit extends ShaUrl{
 	public String message;
 	public boolean distinct;
 	public GitCommitFiles files;
-	public long days;
+	public int days;
 
 	@Override
 	public String toString() {
